window: Update winit to 0.30 and move to main thread
winit has some issues on some platforms when the event loop isn't on the
main thread, so in order to support more platforms in the future - and
to make it easier for users to handle input events from the event loop -
move event loop handling to the main thread instead of being on a
separate thread.

As a side-effect to this and the update to winit-0.30, windows may only
be created during engine initialization now, as window creation needs a
reference to the ActiveEventLoop.

Signed-off-by: James M De Broeck <debroejm@gmail.com>