window: Add method to make the cursor invisible
Cursors are only invisible within the bounds of the affected window. In
order to remove the cursor if the window is focused - as is generally
expected with games - also confine it the window. This happens
differently on different platforms, so try several methods.

Not all platforms automatically show/unconfine the cursor when focus on
the window is lost, so handle this automatically when window focus
changes.

Signed-off-by: James M De Broeck <debroejm@gmail.com>