input: Add priority locking to delegates
It may be the case that a particular layer of your application will need
exclusive access to input without lower layers receiving it. In order to
facilitate this, add a new system to InputDelegates that allows them to
be locked with a priority. If any are locked, only the highest priority
InputDelegate will receive events.

If any InputDelegates are locked, the base-level InputState will also
only return default states (e.g. "unpressed" for keys) when queried.

Signed-off-by: James M De Broeck <debroejm@gmail.com>