render-pass: Add `Result<>` to `build_commands()`
Instead of having every implementation of `build_commands()` swallow errors
with `unwrap()`, change the API to return a
`Result<(), Validated<VulkanoError>>`, which allows implementors to pass along
their errors. Currently, these errors are simply unwrapped at the top-level in
Renderer, but in the future better handling should be applied to them.